Create hello-world slack bot
Uses stack because turns out depndency management is difficult for libs
that haven't been updated for years (which seems to be the case with
these slack API libs). extra-libs-dir/extra-include-dir are required to
work around [an OS X issue][1].

A specific commit of a slack-api library fork is used because the
master/stable version there doesn't work, because a preference field has
been removed from the slack protocol, but the code expects the field to
be there.

[1]: https://github.com/phonohawk/HsOpenSSL/issues/41