Restructure to lib+exec to add tests
Not sure yet if it's a good split between the library and executable
code, but writing tests to it etc. should flesh this out later on. At
least now there is _a_ way to write tests for the code.

Decided to use tasty to be able to write tests in QuickCheck,
SmallCheck, and/or Hspec (the only one used right now).